#139: Merged the QueueBrowsingBeanDefinitionFactory with the JmsBeanDefinitionRegistryPostProcessor
