+++ core/activemq/src/test/java/uk/gov/dwp/queue/triage/core/jms/activemq/spring/JmsBeanDefinitionRegistryPostProcessorTest.java
@@ -3,6 +3,7 @@ package uk.gov.dwp.queue.triage.core.jms.activemq.spring;
 import org.apache.activemq.ActiveMQConnectionFactory;
 import org.junit.After;
 import org.junit.Test;
+import org.springframework.beans.factory.NoSuchBeanDefinitionException;
 import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Bean;
@@ -12,6 +13,7 @@ import org.springframework.core.env.StandardEnvironment;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.jms.listener.DefaultMessageListenerContainer;
 import uk.gov.dwp.queue.triage.core.jms.FailedMessageListener;
+import uk.gov.dwp.queue.triage.core.jms.activemq.browser.QueueBrowserScheduledExecutorService;
 import uk.gov.dwp.queue.triage.core.jms.activemq.configuration.JmsListenerConfig;
 import uk.gov.dwp.queue.triage.core.service.FailedMessageService;
 
@@ -21,6 +23,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 import static org.mockito.Mockito.mock;
+import static uk.gov.dwp.queue.triage.core.jms.activemq.browser.spring.QueueBrowserScheduledExecutorServiceBeanDefinitionFactory.QUEUE_BROWSER_SCHEDULED_EXECUTOR_SERVICE_BEAN_NAME_PREFIX;
 import static uk.gov.dwp.queue.triage.core.jms.activemq.spring.ActiveMQConnectionFactoryBeanDefinitionFactory.ACTIVE_MQ_CONNECTION_FACTORY_BEAN_NAME_PREFIX;
 import static uk.gov.dwp.queue.triage.core.jms.activemq.spring.FailedMessageListenerBeanDefinitionFactory.FAILED_MESSAGE_LISTENER_BEAN_NAME_PREFIX;
 import static uk.gov.dwp.queue.triage.core.jms.activemq.spring.NamedMessageListenerContainerBeanDefinitionFactory.NAMED_MESSAGE_LISTENER_CONTAINER_BEAN_NAME_PREFIX;
@@ -38,36 +41,61 @@ public class JmsBeanDefinitionRegistryPostProcessorTest {
     public void createDefaultMessageListenerContainerForSingleBroker() {
         applicationContext = createApplicationContext("single-broker.yml");
 
-        activeMqConnectionFactoryExistsFor("broker-one");
-        defaultMessageListenerContainerExistsFor("broker-one");
-        failedMessageListenerExistsFor("broker-one");
+        assertThat(activeMqConnectionFactoryBeanFor("read-write-broker"), is(notNullValue(ActiveMQConnectionFactory.class)));
+        assertThat(failedMessageListenerBeanFor("read-write-broker"), is(notNullValue(FailedMessageListener.class)));
+        assertThat(defaultMessageListenerContainerBeanFor("read-write-broker"), is(notNullValue(DefaultMessageListenerContainer.class)));
+
+        assertThat(numberOfBeansOfType(ActiveMQConnectionFactory.class), is(1));
+        assertThat(numberOfBeansOfType(FailedMessageListener.class), is(1));
+        assertThat(numberOfBeansOfType(NamedMessageListenerContainer.class), is(1));
+        assertThat(numberOfBeansOfType(QueueBrowserScheduledExecutorService.class), is(0));
     }
 
     @Test
-    public void createDefaultMessageListenerContainerForMultipleBrokers() {
+    public void createOneReadOnlyAndOneDefaultMessageListenerContainerForMultipleBrokers() {
 
-        applicationContext = createApplicationContext("multiple-brokers.yml");
+        applicationContext = createApplicationContext("multiple-brokers-one-read-only.yml");
 
-        activeMqConnectionFactoryExistsFor("broker-one");
-        defaultMessageListenerContainerExistsFor("broker-one");
-        failedMessageListenerExistsFor("broker-one");
+        assertThat(numberOfBeansOfType(ActiveMQConnectionFactory.class), is(2));
+        assertThat(numberOfBeansOfType(FailedMessageListener.class), is(2));
+        assertThat(numberOfBeansOfType(NamedMessageListenerContainer.class), is(1));
+        assertThat(numberOfBeansOfType(QueueBrowserScheduledExecutorService.class), is(1));
 
-        activeMqConnectionFactoryExistsFor("broker-two");
-        defaultMessageListenerContainerExistsFor("broker-two");
-        failedMessageListenerExistsFor("broker-two");
+        assertThat(activeMqConnectionFactoryBeanFor("read-write-broker"), is(notNullValue(ActiveMQConnectionFactory.class)));
+        assertThat(failedMessageListenerBeanFor("read-write-broker"), is(notNullValue(FailedMessageListener.class)));
+        assertThat(defaultMessageListenerContainerBeanFor("read-write-broker"), is(notNullValue(DefaultMessageListenerContainer.class)));
+
+        assertThat(activeMqConnectionFactoryBeanFor("read-only-broker"), is(notNullValue(ActiveMQConnectionFactory.class)));
+        assertThat(failedMessageListenerBeanFor("read-only-broker"), is(notNullValue(FailedMessageListener.class)));
+        assertThat(queueBrowserScheduledExecutorService("read-only-broker"), is(notNullValue(QueueBrowserScheduledExecutorService.class))); 
     }
 
-    private void failedMessageListenerExistsFor(String brokerName) {
-        applicationContext.getBean(FAILED_MESSAGE_LISTENER_BEAN_NAME_PREFIX + brokerName, FailedMessageListener.class);
+    private FailedMessageListener failedMessageListenerBeanFor(String brokerName) {
+        return getBean(FAILED_MESSAGE_LISTENER_BEAN_NAME_PREFIX + brokerName, FailedMessageListener.class);
     }
 
-    private void defaultMessageListenerContainerExistsFor(String brokerName) {
-        applicationContext.getBean(NAMED_MESSAGE_LISTENER_CONTAINER_BEAN_NAME_PREFIX + brokerName, DefaultMessageListenerContainer.class);
+    private DefaultMessageListenerContainer defaultMessageListenerContainerBeanFor(String brokerName) {
+        return getBean(NAMED_MESSAGE_LISTENER_CONTAINER_BEAN_NAME_PREFIX + brokerName, DefaultMessageListenerContainer.class);
     }
 
-    private void activeMqConnectionFactoryExistsFor(String brokerName) {
-        ActiveMQConnectionFactory activeMQConnectionFactory = applicationContext.getBean(ACTIVE_MQ_CONNECTION_FACTORY_BEAN_NAME_PREFIX + brokerName, ActiveMQConnectionFactory.class);
-        assertThat(activeMQConnectionFactory, is(notNullValue(ActiveMQConnectionFactory.class)));
+    private ActiveMQConnectionFactory activeMqConnectionFactoryBeanFor(String brokerName) {
+        return getBean(ACTIVE_MQ_CONNECTION_FACTORY_BEAN_NAME_PREFIX + brokerName, ActiveMQConnectionFactory.class);
+    }
+
+    private QueueBrowserScheduledExecutorService queueBrowserScheduledExecutorService(String brokerName) {
+        return getBean(QUEUE_BROWSER_SCHEDULED_EXECUTOR_SERVICE_BEAN_NAME_PREFIX + brokerName, QueueBrowserScheduledExecutorService.class);
+    }
+
+    private <T> T getBean(String beanName, Class<T> beanType) {
+        try {
+            return applicationContext.getBean(beanName, beanType);
+        } catch (NoSuchBeanDefinitionException e) {
+            return null;
+        }
+    }
+
+    private int numberOfBeansOfType(Class<?> beanType) {
+        return applicationContext.getBeanNamesForType(beanType).length;
     }
 
     private AnnotationConfigApplicationContext createApplicationContext(String yamlFilename) {
@@ -93,20 +121,10 @@ public class JmsBeanDefinitionRegistryPostProcessorTest {
     @Configuration
     public static class AdditionalConfig {
 
-        static String yamlFile = "application-single-broker.yml";
-
         @Bean
         public FailedMessageService failedMessageService() {
             return mock(FailedMessageService.class);
         }
 
-//        @Bean
-//        @Order(Ordered.HIGHEST_PRECEDENCE)
-//        public static PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer() {
-//
-//            PropertySourcesPlaceholderConfigurer propertySourcesPlaceholderConfigurer = new PropertySourcesPlaceholderConfigurer();
-//            propertySourcesPlaceholderConfigurer.setProperties(yamlPropertiesFactoryBean.getObject());
-//            return propertySourcesPlaceholderConfigurer;
-//        }
     }
 }
