--- core/activemq/src/test/resources/single-broker-read-only.yml
@@ -1,11 +1,0 @@
-jms:
-  activemq:
-    brokers:
-      - name: broker-one
-        url: tcp://localhost:1111
-        queue: ActiveMQ.DLQ
-        readOnly: true
-        read:
-          frequency: 500
-          timeUnit: MILLISECOND
-
